requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-chroma
langchain-huggingface
pypdf
google-generativeai